+++ docs/changelog.md
@@ -1,5 +1,11 @@
 # 更新日志
 
+## v3.0.0
+
+- 新增 windows/linux 换行符转换 `FileEndingUtils`
+- 移除 `borax.loader`
+- 移除 `borax.decorators.admin`
+
 ## v1.4.2 (20190717)
 
 - `counters.serials` 模块